backend/requirements.txt
flask
flask-cors
firebase-admin
python-dotenv
gunicorn
psycopg2-binary
sqlalchemy
flask-sqlalchemy
redis
flask-caching